democompositor: Add parsing of entries and testcode
Implement parsing of the json definition. Initially be strict to
parse the file correctly. Add some basic tests for checking how
data and file parsing is working and feed some invalid input into
the parser and check the result.

Use macro for checking the JSON type and doing the conversion to
allow an early return and to avoid having to write the same code
over and over again.